Rename views to follow new naming
